ProductImageController is done (part 02)

+++ src/main/java/ca/com/rlsp/ecommerce/controller/ProductImageController.java
@@ -2,7 +2,6 @@ package ca.com.rlsp.ecommerce.controller;
 
 import ca.com.rlsp.ecommerce.model.ProductImage;
 import ca.com.rlsp.ecommerce.model.dto.ProductImageDTO;
-import ca.com.rlsp.ecommerce.repository.ProductImageRepository;
 import ca.com.rlsp.ecommerce.service.ProductImageService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -30,6 +29,8 @@ public class ProductImageController {
 
         productImage = productImageService.saveAndFlush(productImage);
 
+
+        // Use a DTO to avoid a recursive response
         ProductImageDTO productImageDTO = new ProductImageDTO();
         productImageDTO.setId(productImage.getId());
         productImageDTO.setEcommerceCompany(productImage.getEcommerceCompany().getId());
@@ -44,9 +45,12 @@ public class ProductImageController {
 
 
     @ResponseBody
-    @DeleteMapping(value = "**/deleteAllProductImages/{productId}")
+    @DeleteMapping(value = "/deleteAllProductImages/{productId}")
     public ResponseEntity<?> deleteAllProductImages(@PathVariable("productId") Long productId) {
 
+        if(!productImageService.isImageSavedByProductIdIntoDB(productId)) {
+            return new ResponseEntity<String>(IMAGE_ALREADY_REMOVED_OR_NOT_SAVED_BY_ID + productId, HttpStatus.OK);
+        }
         productImageService.deleteAllProductImages(productId);
 
         return new ResponseEntity<String>(ALL_PRODUCT_IMAGES_WERE_REMOVED, HttpStatus.NO_CONTENT);
@@ -58,7 +62,7 @@ public class ProductImageController {
     @DeleteMapping(value = "/deleteProductImage")
     public ResponseEntity<?> deleteProductImageByObject(@RequestBody ProductImage productImage) {
 
-        if(!productImageService.existsImageSavedById(productImage.getId())) {
+        if(!productImageService.isImageSavedByProductImageIdIntoDB(productImage.getId())) {
             return new ResponseEntity<String>(IMAGE_ALREADY_REMOVED_OR_NOT_SAVED_BY_ID + productImage.getId(), HttpStatus.OK);
         }
 
@@ -73,7 +77,7 @@ public class ProductImageController {
     @DeleteMapping(value = "/deleteProductImageById/{productImageId}")
     public ResponseEntity<?> deleteProductImageById(@PathVariable("productImageId") Long id) {
 
-        if(!productImageService.existsImageSavedById(id)) {
+        if(!productImageService.isImageSavedByProductImageIdIntoDB(id)) {
             return new ResponseEntity<String>(IMAGE_ALREADY_REMOVED_OR_NOT_SAVED_BY_ID + id, HttpStatus.OK);
         }
 
@@ -83,12 +87,16 @@ public class ProductImageController {
     }
 
     @ResponseBody
-    @GetMapping(value = "**/getProductImageByProduct/{productId}")
-    public ResponseEntity<List<ProductImageDTO>> getProductImageByProduct(@PathVariable("productId") Long productId){
+    @GetMapping(value = "/getProductImageByProduct/{productId}")
+    public ResponseEntity<?> getProductImageByProduct(@PathVariable("productId") Long productId){
+
+        if(!productImageService.isImageSavedByProductIdIntoDB(productId)) {
+            return new ResponseEntity<String>(IMAGE_ALREADY_REMOVED_OR_NOT_SAVED_BY_ID + productId, HttpStatus.OK);
+        }
 
         List<ProductImageDTO> dtos = new ArrayList<ProductImageDTO>();
 
-        List<ProductImage> productImages = productImageService.fetchProductImages(productId);
+        List<ProductImage> productImages = productImageService.fetchAllProductImagesByProduct(productId);
 
         for (ProductImage productImage : productImages) {
